# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MediBook Pro - Clinic & Hospital Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/suggest-clinics:
    post:
      operationId: suggestClinics
      tags: [clinics]
      summary: AI-powered clinic suggestions based on symptoms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestClinicsBody"
      responses:
        "200":
          description: Suggested clinics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestClinicsResponse"

  # ADMIN
...
  /admin/clinics:
    get:
      operationId: adminListClinics
      tags: [admin]
      summary: Admin list all clinics with filters
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: All clinics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Clinic"
...
  /admin/revenue:
    get:
      operationId: getAdminRevenue
      tags: [admin]
      summary: Get platform revenue analytics
      parameters:
        - name: period
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Revenue data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueReport"
  /admin/complaints:
    get:
      operationId: listComplaints
      tags: [admin]
      summary: List all complaints
      responses:
        "200":
          description: Complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"

    post:
      operationId: createComplaint
      tags: [admin]
      summary: Submit a complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplaintBody"
      responses:
        "201":
          description: Complaint submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /admin/complaints/{id}:
    patch:
      operationId: updateComplaint
      tags: [admin]
      summary: Update complaint status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComplaintBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /analytics/bookings:
    get:
      operationId: getBookingAnalytics
      tags: [analytics]
      summary: Get booking analytics over time
      parameters:
        - name: clinicId
          in: query
          schema:
            type: ["integer", "null"]
        - name: period
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Booking analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingAnalytics"
components:
  schemas:
...
    Clinic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        ownerName:
          type: string
        ownerId:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        logoUrl:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        subscriptionPlan:
          type: string
          enum: [basic, premium, enterprise]
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        isEmergencyAvailable:
          type: boolean
        workingHours:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - ownerName
        - ownerId
        - categoryId
        - email
        - phone
        - address
        - city
        - state
        - pincode
        - subscriptionPlan
        - status
        - reviewCount
        - isEmergencyAvailable
        - createdAt
        - updatedAt
...
    MonthlyRevenueStat:
      type: object
      properties:
        month:
          type: string
        revenue:
          type: number
        appointments:
          type: integer
      required:
        - month
        - revenue
        - appointments

    # DOCTOR SCHEMAS
...
    SuggestClinicsBody:
      type: object
      properties:
        symptom:
          type: string
        city:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
      required:
        - symptom
    SuggestClinicsResponse:
      type: object
      properties:
        category:
          type: string
        reason:
          type: string
        clinics:
          type: array
          items:
            $ref: "#/components/schemas/Clinic"
      required:
        - category
        - reason
        - clinics

    # ADMIN SCHEMAS
...
    RevenuePlanStat:
      type: object
      properties:
        plan:
          type: string
        count:
          type: integer
        revenue:
          type: number
      required:
        - plan
        - count
        - revenue
    RevenueReport:
      type: object
      properties:
        totalRevenue:
          type: number
        monthlyData:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyRevenueStat"
        byPlan:
          type: array
          items:
            $ref: "#/components/schemas/RevenuePlanStat"
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryRevenueStat"
      required:
        - totalRevenue
        - monthlyData
        - byPlan
        - byCategory
    CategoryRevenueStat:
      type: object
      properties:
        category:
          type: string
        revenue:
          type: number
        count:
          type: integer
      required:
        - category
        - revenue
        - count
    Complaint:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        clinicId:
          type: ["integer", "null"]
        subject:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - subject
        - description
        - status
        - priority
        - createdAt
        - updatedAt
    CreateComplaintBody:
      type: object
      properties:
        userId:
          type: ["integer", "null"]
        clinicId:
          type: ["integer", "null"]
        subject:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
      required:
        - subject
        - description
        - priority
    UpdateComplaintBody:
      type: object
      properties:
        status:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
    BookingAnalytics:
      type: object
      properties:
        dailyBookings:
          type: array
          items:
            $ref: "#/components/schemas/DailyBookingStat"
        byStatus:
          type: array
          items:
            $ref: "#/components/schemas/StatusStat"
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryStat"
      required:
        - dailyBookings
        - byStatus
        - byCategory
    DailyBookingStat:
      type: object
      properties:
        date:
          type: string
        count:
          type: integer
      required:
        - date
        - count
    StatusStat:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
      required:
        - status
        - count
    CategoryStat:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
      required:
        - category
        - count