# lib/db/src/schema/clinics.ts
import { pgTable, text, serial, timestamp, boolean, integer, numeric } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const clinicsTable = pgTable("clinics", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  ownerName: text("owner_name").notNull(),
  ownerId: integer("owner_id").notNull(),
  categoryId: integer("category_id").notNull(),
  email: text("email").notNull(),
  phone: text("phone").notNull(),
  address: text("address").notNull(),
  city: text("city").notNull(),
  state: text("state").notNull(),
  pincode: text("pincode").notNull(),
  latitude: numeric("latitude", { precision: 10, scale: 7 }),
  longitude: numeric("longitude", { precision: 10, scale: 7 }),
  logoUrl: text("logo_url"),
  photos: text("photos").array().default([]),
  services: text("services").array().default([]),
  whatsappNumber: text("whatsapp_number"),
  instagramUrl: text("instagram_url"),
  websiteUrl: text("website_url"),
  subscriptionPlan: text("subscription_plan").notNull().default("basic"),
  status: text("status").notNull().default("pending"),
  rating: numeric("rating", { precision: 3, scale: 2 }),
  reviewCount: integer("review_count").notNull().default(0),
  isEmergencyAvailable: boolean("is_emergency_available").notNull().default(false),
  workingHours: text("working_hours"),
  licenseUrl: text("license_url"),
  rejectionReason: text("rejection_reason"),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertClinicSchema = createInsertSchema(clinicsTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertClinic = z.infer<typeof insertClinicSchema>;
export type Clinic = typeof clinicsTable.$inferSelect;